MAE-575: Sets the num_terms to 1 for memberships with manually entered dates
If we choose to enter dates manually, we have to add the membership fee field
to provide the price. Otherwise the membership will be free according to
the default behaviour of webform_civicrm.

We don't want to enable the membership fee field so that default membership
fee from CiviCRM membership type settings will be loaded. But if we
didn't create the membership will be free and no contribution will be
created.

After the investigation, I found that `webform_civicrm` sets the num_terms
to 0 for manually entered dates memberships so the quantity and the total
price of the line item will be zero.

This commit sets the number of terms to 1 for memberships with manually entered dates.