IO-341: Fix the prorated calculation if the membership select page was the last
The membership types must be fixed to do the prorated calculation and
the validate callbacks will not run at the confirmation page. It was
reported that we need to go back and forward to activate the prorated
calculation. Here, we delay the check for membership types so that we
can always (if the webform configuration passed our remaining checks
about the number of terms and the membership fee is not set manually)
inject the validate callbacks early so that we can do the prorated
calculation at the confirmation page.